Guest 5 用户数据 - 收藏App  ok

+++ Gate/VO/src/main/java/xyz/spc/gate/vo/Guest/datas/CollectCountVO.java
@@ -12,9 +12,9 @@ import xyz.spc.gate.vo.BaseVO;
 @AllArgsConstructor
 public class CollectCountVO extends BaseVO {
 
-    private Long postCount;
+    private Integer postCount;
 
-    private Long fileCount;
+    private Integer fileCount;
 
-    private Long clusterCount;
+    private Integer clusterCount;
 }
